+++ src/Models/Comment.php
@@ -2,9 +2,6 @@
 
 namespace Firefly\FilamentBlog\Models;
 
-use Filament\Forms\Components\Select;
-use Filament\Forms\Components\Textarea;
-use Filament\Forms\Components\Toggle;
 use Firefly\FilamentBlog\Database\Factories\CommentFactory;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
@@ -59,23 +56,6 @@ class Comment extends Model
     public function scopeApproved(Builder $query)
     {
         return $query->where('approved', true);
-    }
-
-    public static function getForm(): array
-    {
-        return [
-            Select::make('user_id')
-                ->relationship('user', config('filamentblog.user.columns.name'))
-                ->required(),
-            Select::make('post_id')
-                ->relationship('post', 'title')
-                ->required(),
-            Textarea::make('comment')
-                ->required()
-                ->maxLength(65535)
-                ->columnSpanFull(),
-            Toggle::make('approved'),
-        ];
     }
 
     public function getTable()
