--- src/Resources/CommentResource/Pages/CreateComment.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace Firefly\FilamentBlog\Resources\CommentResource\Pages;
-
-use Filament\Resources\Pages\CreateRecord;
-use Firefly\FilamentBlog\Resources\CommentResource;
-
-class CreateComment extends CreateRecord
-{
-    protected static string $resource = CommentResource::class;
-}
