--- src/Resources/CommentResource/Pages/EditComment.php
@@ -1,21 +1,0 @@
-<?php
-
-namespace Firefly\FilamentBlog\Resources\CommentResource\Pages;
-
-use Filament\Actions\DeleteAction;
-use Filament\Actions;
-use Filament\Resources\Pages\EditRecord;
-use Firefly\FilamentBlog\Resources\CommentResource;
-
-class EditComment extends EditRecord
-{
-    protected static string $resource = CommentResource::class;
-
-    protected function getHeaderActions(): array
-    {
-        return [
-            DeleteAction::make(),
-        ];
-    }
-
-}
